+++ src/main/java/cart/dao/ProductDao.java
@@ -1,4 +1,4 @@
-package cart.service;
+package cart.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
