+++ src/main/java/cart/dao/ProductEntity.java
@@ -1,14 +1,20 @@
-package cart.controller;
+package cart.dao;
 
-public class ProductSearchResponse {
+public class ProductEntity {
 
-    private final Long id;
+    private Long id;
     private final String name;
     private final int price;
     private final String imageUrl;
 
-    public ProductSearchResponse(final Long id, final String name, final int price, final String imageUrl) {
+    public ProductEntity(final Long id, final String name, final int price, final String imageUrl) {
         this.id = id;
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public ProductEntity(final String name, final int price, final String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
