+++ src/main/java/cart/service/dto/ProductModifyRequest.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.service.dto;
 
 public class ProductModifyRequest {
