+++ src/main/java/cart/service/dto/ProductSearchResponse.java
@@ -1,20 +1,14 @@
-package cart.service;
+package cart.service.dto;
 
-public class ProductEntity {
+public class ProductSearchResponse {
 
-    private Long id;
+    private final Long id;
     private final String name;
     private final int price;
     private final String imageUrl;
 
-    public ProductEntity(final Long id, final String name, final int price, final String imageUrl) {
+    public ProductSearchResponse(final Long id, final String name, final int price, final String imageUrl) {
         this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public ProductEntity(final String name, final int price, final String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
