+++ src/test/java/cart/dao/ProductDaoTest.java
@@ -1,5 +1,7 @@
-package cart.service;
+package cart.dao;
 
+import cart.dao.ProductDao;
+import cart.dao.ProductEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
